requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
atomicwrites==1.4.1
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
contourpy==1.0.6
cycler==0.11.0
exchangerates==0.3.4
fonttools==4.38.0
greenlet==1.1.2
idna==3.3
iniconfig==1.1.1
kiwisolver==1.4.4
lxml==4.9.1
matplotlib==3.6.2
minio==7.1.12
numpy==1.23.4
packaging==21.3
pika==1.3.1
Pillow==9.3.0
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
response==0.5.0
scipy==1.9.3
six==1.16.0
SQLAlchemy==1.4.40
SQLAlchemy-Utils==0.38.3
tomli==2.0.1
urllib3==1.26.11
